+++ src/app/my/profile/page.tsx
@@ -1,4 +1,4 @@
-import ProfileForm from "@/components/my/ProfileForm";
+import ProfileForm from "@/containers/my/ProfileForm";
 import { getMe } from "@/lib/apis/user/getMe";
 
 const Page = async () => {
